+++ CHANGELOG.md
@@ -1,3 +1,12 @@
-## [0.0.1] - TODO: Add release date.
+## [1.0.0] - Stable release
 
-* TODO: Describe initial release.
+* Full rewrite of the library.
+* You can provide a custom builder to build the character entities.
+* Provided a wide range of pre-build builders to use (circles or rectangles to be used in light or dark mode).
+* Example added.
+* Readme edited. Installtion does _not_ have to be described there, that's what pub's "Install" tab is for.
+* License added.
+
+## [0.0.1] - Initial release
+
+* Initial release.